:building_construction: Combine all examples into examples/Eg01.basic.
