+++ .travis.yml
@@ -6,9 +6,4 @@ rust:
 script:
 - cargo build --manifest-path ./hakurei/Cargo.toml
 - cargo build --manifest-path ./hakurei-macros/Cargo.toml
-- cargo build --manifest-path ./examples/Eg001.triangle/Cargo.toml
-- cargo build --manifest-path ./examples/Eg002.index_buffer/Cargo.toml
-- cargo build --manifest-path ./examples/Eg003.staging_buffer/Cargo.toml
-- cargo build --manifest-path ./examples/Eg004.uniform_buffer/Cargo.toml
-- cargo build --manifest-path ./examples/Eg005.texture_mapping/Cargo.toml
-- cargo build --manifest-path ./examples/Eg006.box/Cargo.toml
+- cargo build --manifest-path ./examples/Eg01.basic/Cargo.toml
