--- examples/Eg002.index_buffer/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "index_buffer"
-version = "0.1.0"
-authors = ["usami-renko <usami-ssc@protonmail.com>"]
-description = "Use index buffer to draw an rectangle."
-
-[dependencies]
-hakurei = { path = "../../hakurei" }
-hakurei-macros = { path = "../../hakurei-macros" }
