--- examples/Eg003.staging_buffer/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "staging_buffer"
-version = "0.1.0"
-authors = ["usami-renko <usami-ssc@protonmail.com>"]
-description = "Use staging buffer to transfer vertex data to GPU."
-
-[dependencies]
-hakurei = { path = "../../hakurei" }
-hakurei-macros = { path = "../../hakurei-macros" }
