--- examples/Eg004.uniform_buffer/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "uniform_buffer"
-version = "0.1.0"
-authors = ["usami-renko <usami-ssc@protonmail.com>"]
-description = "Use uniform buffer in rendering."
-
-[dependencies]
-hakurei = { path = "../../hakurei" }
-hakurei-macros = { path = "../../hakurei-macros" }
-cgmath = "0.16.1"
