+++ examples/Eg01.basic/run-all.sh
@@ -1,0 +1,7 @@
+
+cargo run --bin 01
+cargo run --bin 02
+cargo run --bin 03
+cargo run --bin 04
+cargo run --bin 05
+cargo run --bin 06
