+++ examples/Eg01.basic/shaders/box.frag.spv
