+++ examples/Eg01.basic/shaders/box.vert.spv
