+++ examples/Eg01.basic/shaders/index.frag.spv
