+++ examples/Eg01.basic/shaders/index.vert.spv
