+++ examples/Eg01.basic/shaders/staging.frag.spv
