+++ examples/Eg01.basic/shaders/staging.vert.spv
