+++ examples/Eg01.basic/shaders/texture.frag.spv
