+++ examples/Eg01.basic/shaders/texture.vert.spv
