+++ examples/Eg01.basic/shaders/triangle.frag.spv
