+++ examples/Eg01.basic/shaders/triangle.vert.spv
