+++ examples/Eg01.basic/shaders/uniform.frag.spv
