+++ examples/Eg01.basic/shaders/uniform.vert.spv
