+++ examples/Eg01.basic/src/01.triangle/triangle.frag
