+++ examples/Eg01.basic/src/03.staging/main.rs
@@ -1,12 +1,7 @@
 
-mod data;
-
-pub use self::data::{ Vertex, UboObject };
-
 #[macro_use]
 extern crate hakurei_macros;
 extern crate hakurei;
-extern crate cgmath;
 
 use hakurei::prelude::*;
 use hakurei::prelude::config::*;
@@ -15,96 +10,69 @@ use hakurei::prelude::pipeline::*;
 use hakurei::prelude::resources::*;
 use hakurei::prelude::sync::*;
 use hakurei::prelude::input::*;
-use hakurei::prelude::utility::*;
-
-use cgmath::{ Matrix4, SquareMatrix, Point3 };
 
 use std::path::Path;
 
-const WINDOW_TITLE: &'static str = "Box Example";
+const WINDOW_TITLE: &'static str = "Staging Buffer Example";
 const WINDOW_WIDTH:  u32 = 800;
 const WINDOW_HEIGHT: u32 = 600;
+const VERTEX_SHADER_PATH  : &'static str = "shaders/staging.vert.spv";
+const FRAGMENT_SHADER_PATH: &'static str = "shaders/staging.frag.spv";
 
-struct BoxProcedure {
+define_input! {
+    #[binding = 0, rate = vertex]
+    struct Vertex {
+        #[location = 0, format = vec2]
+        pos:   [f32; 2],
+        #[location = 1, format = vec4]
+        color: [f32; 4],
+    }
+}
 
-    vertex_data: Vec<Vertex>,
-    index_data : Vec<uint32_t>,
+const VERTEX_DATA: [Vertex; 3] = [
+    Vertex { pos: [ 0.0, -0.5], color: [1.0, 0.0, 0.0, 1.0], },
+    Vertex { pos: [ 0.5,  0.5], color: [0.0, 1.0, 0.0, 1.0], },
+    Vertex { pos: [-0.5,  0.5], color: [0.0, 0.0, 1.0, 1.0], },
+];
 
-    buffer_storage: HaBufferRepository,
-    vertex_item   : BufferSubItem,
-    index_item    : BufferSubItem,
+struct StagingBufferProcedure {
+
+    vertex_data  : Vec<Vertex>,
+    vertex_buffer: HaBufferRepository,
+    vertex_item  : BufferSubItem,
 
     graphics_pipeline: HaGraphicsPipeline,
-
-    ubo_data   : Vec<UboObject>,
-    ubo_buffer : HaBufferRepository,
-    ubo_item   : BufferSubItem,
-    ubo_storage: HaDescriptorRepository,
-    ubo_set    : DescriptorSetItem,
 
     command_pool   : HaCommandPool,
     command_buffers: Vec<HaCommandBuffer>,
 
-    camera: HaStageCamera,
-
     present_availables: Vec<HaSemaphore>,
 }
 
-impl BoxProcedure {
+impl StagingBufferProcedure {
 
-    fn new() -> BoxProcedure {
-        let camera = CameraConfigurator::config()
-            .place_at(Point3::new(0.0, 0.0, 3.0))
-            .screen_dimension(WINDOW_WIDTH, WINDOW_HEIGHT)
-            .for_stage_camera();
-
-        BoxProcedure {
-            vertex_data: data::VERTEX_DATA.to_vec(),
-            index_data : data::INDEX_DATA.to_vec(),
-
-            buffer_storage: HaBufferRepository::empty(),
-            vertex_item   : BufferSubItem::unset(),
-            index_item    : BufferSubItem::unset(),
+    fn new() -> StagingBufferProcedure {
+        StagingBufferProcedure {
+            vertex_data  : VERTEX_DATA.to_vec(),
+            vertex_buffer: HaBufferRepository::empty(),
+            vertex_item  : BufferSubItem::unset(),
 
             graphics_pipeline: HaGraphicsPipeline::uninitialize(),
-
-            ubo_data: vec![
-                UboObject {
-                    projection: camera.proj_matrix(),
-                    view      : camera.view_matrix(),
-                    model     : Matrix4::identity(),
-                },
-            ],
-            ubo_buffer : HaBufferRepository::empty(),
-            ubo_item   : BufferSubItem::unset(),
-            ubo_storage: HaDescriptorRepository::empty(),
-            ubo_set: DescriptorSetItem::unset(),
 
             command_pool: HaCommandPool::uninitialize(),
             command_buffers: vec![],
 
-            camera,
-
             present_availables: vec![],
         }
     }
-
-    fn update_uniforms(&mut self, device: &HaLogicalDevice) -> Result<(), ProcedureError> {
-
-        self.ubo_data[0].model = self.camera.object_model_transformation();
-        self.ubo_data[0].view  = self.camera.view_matrix();
-        self.ubo_buffer.tranfer_data(device, &self.ubo_data, &self.ubo_item)?;
-
-        Ok(())
-    }
 }
 
-impl ProgramProc for BoxProcedure {
+impl ProgramProc for StagingBufferProcedure {
 
     fn assets(&mut self, device: &HaLogicalDevice, generator: &ResourceGenerator) -> Result<(), ProcedureError> {
 
-        // staging buffer
-        let staging_buffer_config = BufferConfig::init(
+        // vertex buffer
+        let mut staging_buffer_config = BufferConfig::init(
             &[BufferUsageFlag::TransferSrcBit],
             &[
                 MemoryPropertyFlag::HostVisibleBit,
@@ -112,22 +80,8 @@ impl ProgramProc for BoxProcedure {
             ],
             &[]
         );
+        let _ = staging_buffer_config.add_item(data_size!(self.vertex_data, Vertex));
 
-        let mut vertex_staging_buffer = staging_buffer_config.clone();
-        let mut index_staging_buffer  = staging_buffer_config.clone();
-
-        let _ = vertex_staging_buffer.add_item(data_size!(self.vertex_data, Vertex));
-        let _ = index_staging_buffer.add_item(data_size!(self.index_data, uint32_t));
-
-        let mut buffer_allocator = generator.buffer();
-        let vertex_staging_item = buffer_allocator.attach_buffer(vertex_staging_buffer)?.pop().unwrap();
-        let index_staging_item  = buffer_allocator.attach_buffer(index_staging_buffer )?.pop().unwrap();
-
-        let mut staging_repository = buffer_allocator.allocate()?;
-        staging_repository.tranfer_data(device, &self.vertex_data, &vertex_staging_item)?;
-        staging_repository.tranfer_data(device, &self.index_data,  &index_staging_item)?;
-
-        // vertex, index buffer
         let mut vertex_buffer_config = BufferConfig::init(
             &[
                 BufferUsageFlag::TransferDstBit,
@@ -138,61 +92,20 @@ impl ProgramProc for BoxProcedure {
         );
         let _ = vertex_buffer_config.add_item(data_size!(self.vertex_data, Vertex));
 
-        let mut index_buffer_config = BufferConfig::init(
-            &[
-                BufferUsageFlag::TransferDstBit,
-                BufferUsageFlag::IndexBufferBit,
-            ],
-            &[MemoryPropertyFlag::DeviceLocalBit],
-            &[]
-        );
-        let _ = index_buffer_config.add_item(data_size!(self.index_data, uint32_t));
+        let mut staging_allocator = generator.buffer();
+        let stage_buffer_item = staging_allocator.attach_buffer(staging_buffer_config)?.pop().unwrap();
 
-        buffer_allocator.reset();
-        self.vertex_item = buffer_allocator.attach_buffer(vertex_buffer_config)?.pop().unwrap();
-        self.index_item  = buffer_allocator.attach_buffer(index_buffer_config )?.pop().unwrap();
+        let mut staging_repository = staging_allocator.allocate()?;
+        staging_repository.tranfer_data(device, &self.vertex_data, &stage_buffer_item)?;
 
-        self.buffer_storage = buffer_allocator.allocate()?;
-        self.buffer_storage.copy_buffer_to_buffer(device, &vertex_staging_item, &self.vertex_item)?;
-        self.buffer_storage.copy_buffer_to_buffer(device, &index_staging_item,  &self.index_item)?;
+        let mut vertex_allocator = generator.buffer();
+        self.vertex_item = vertex_allocator.attach_buffer(vertex_buffer_config)?.pop().unwrap();
+        self.vertex_buffer = vertex_allocator.allocate()?;
+        self.vertex_buffer.copy_buffer_to_buffer(
+            device,
+            &stage_buffer_item,
+            &self.vertex_item)?;
         staging_repository.cleanup(device);
-
-        // uniform buffer
-        let mut uniform_buffer_config = BufferConfig::init(
-            &[BufferUsageFlag::UniformBufferBit],
-            &[
-                MemoryPropertyFlag::HostVisibleBit,
-                MemoryPropertyFlag::HostCoherentBit,
-            ],
-            &[],
-        );
-        let _ = uniform_buffer_config.add_item(data_size!(self.ubo_data, UboObject));
-
-        let mut uniform_allocator = generator.buffer();
-        self.ubo_item = uniform_allocator.attach_buffer(uniform_buffer_config)?.pop().unwrap();
-        self.ubo_buffer = uniform_allocator.allocate()?;
-        self.ubo_buffer.tranfer_data(device, &self.ubo_data, &self.ubo_item)?;
-
-        // descriptor
-        let ubo_info = DescriptorBufferBindingInfo {
-            binding: 0,
-            type_: BufferDescriptorType::UniformBuffer,
-            count: 1,
-            element_size: data_size!(self.ubo_data, UboObject),
-            buffer: self.ubo_item.clone(),
-        };
-        let mut descriptor_set_config = DescriptorSetConfig::init(&[]);
-        let ubo_binding_index = descriptor_set_config.add_buffer_binding(ubo_info, &[
-            ShaderStageFlag::VertexStage,
-        ]);
-
-        let mut descriptor_allocator = generator.descriptor(&[]);
-        let (descriptor_set_item, descriptor_binding_items) = descriptor_allocator.attach_descriptor_set(descriptor_set_config);
-        let ubo_descriptor_item = descriptor_binding_items[ubo_binding_index].clone();
-
-        self.ubo_storage = descriptor_allocator.allocate()?;
-        self.ubo_storage.update_descriptors(device, &[ubo_descriptor_item]);
-        self.ubo_set = descriptor_set_item;
 
         Ok(())
     }
@@ -201,11 +114,11 @@ impl ProgramProc for BoxProcedure {
         // shaders
         let vertex_shader = HaShaderInfo::setup(
             ShaderStageFlag::VertexStage,
-            Path::new("shaders/box.vert.spv"),
+            Path::new(VERTEX_SHADER_PATH),
             None);
         let fragment_shader = HaShaderInfo::setup(
             ShaderStageFlag::FragmentStage,
-            Path::new("shaders/box.frag.spv"),
+            Path::new(FRAGMENT_SHADER_PATH),
             None);
         let shader_infos = vec![
             vertex_shader,
@@ -230,10 +143,8 @@ impl ProgramProc for BoxProcedure {
 
         let render_pass = render_pass_builder.build(device, swapchain)?;
         let viewport = HaViewport::setup(swapchain.extent);
-
         let pipeline_config = GraphicsPipelineConfig::new(shader_infos, vertex_input_desc, render_pass)
             .setup_viewport(viewport)
-            .add_descriptor_set(self.ubo_storage.set_layout_at(&self.ubo_set))
             .finish_config();
 
         let mut pipeline_builder = GraphicsPipelineBuilder::init();
@@ -246,13 +157,11 @@ impl ProgramProc for BoxProcedure {
     }
 
     fn subresources(&mut self, device: &HaLogicalDevice) -> Result<(), ProcedureError> {
-
         // sync
         for _ in 0..self.graphics_pipeline.frame_count() {
             let present_available = HaSemaphore::setup(device)?;
             self.present_availables.push(present_available);
         }
-
         Ok(())
     }
 
@@ -270,10 +179,8 @@ impl ProgramProc for BoxProcedure {
             recorder.begin_record(&[CommandBufferUsageFlag::SimultaneousUseBit])?
                 .begin_render_pass(&self.graphics_pipeline, frame_index)
                 .bind_pipeline(&self.graphics_pipeline)
-                .bind_vertex_buffers(0, &self.buffer_storage.vertex_binding_infos(&[&self.vertex_item]))
-                .bind_index_buffers(&self.buffer_storage.index_binding_info(&self.index_item))
-                .bind_descriptor_sets(&self.graphics_pipeline, 0, &self.ubo_storage.descriptor_binding_infos(&[&self.ubo_set]))
-                .draw_indexed(self.index_data.len() as uint32_t, 1, 0, 0, 0)
+                .bind_vertex_buffers(0, &self.vertex_buffer.vertex_binding_infos(&[&self.vertex_item]))
+                .draw(self.vertex_data.len() as uint32_t, 1, 0, 0)
                 .end_render_pass()
                 .finish()?;
         }
@@ -284,8 +191,6 @@ impl ProgramProc for BoxProcedure {
     }
 
     fn draw(&mut self, device: &HaLogicalDevice, device_available: &HaFence, image_available: &HaSemaphore, image_index: usize, _: f32) -> Result<&HaSemaphore, ProcedureError> {
-
-        self.update_uniforms(device)?;
 
         let submit_infos = [
             QueueSubmitBundle {
@@ -323,18 +228,14 @@ impl ProgramProc for BoxProcedure {
 
         self.graphics_pipeline.cleanup(device);
         self.command_pool.cleanup(device);
-        self.ubo_storage.cleanup(device);
-        self.ubo_buffer.cleanup(device);
-        self.buffer_storage.cleanup(device);
+        self.vertex_buffer.cleanup(device);
     }
 
-    fn react_input(&mut self, inputer: &ActionNerve, delta_time: f32) -> SceneAction {
+    fn react_input(&mut self, inputer: &ActionNerve, _: f32) -> SceneAction {
 
         if inputer.is_key_pressed(HaKeycode::Escape) {
             return SceneAction::Terminal
         }
-
-        self.camera.react_input(inputer, delta_time);
 
         SceneAction::Rendering
     }
@@ -342,7 +243,7 @@ impl ProgramProc for BoxProcedure {
 
 fn main() {
 
-    let procecure = BoxProcedure::new();
+    let procecure = StagingBufferProcedure::new();
     let mut config = EngineConfig::default();
     config.window.dimension = Dimension2D {
         width : WINDOW_WIDTH,
