+++ examples/Eg01.basic/src/03.staging/staging.vert
