+++ examples/Eg01.basic/src/04.uniform/uniform.vert
