+++ examples/Eg01.basic/src/05.texture/texture.vert
