+++ examples/Eg01.basic/src/06.box/box.frag
