+++ examples/Eg01.basic/src/06.box/data.rs
