+++ examples/Eg01.basic/src/lib.rs
@@ -1,0 +1,1 @@
+
