--- examples/run-all.sh
@@ -1,14 +1,0 @@
-
-cd Eg001.triangle
-cargo run
-cd ../Eg002.index_buffer
-cargo run
-cd ../Eg003.staging_buffer
-cargo run
-cd ../Eg004.uniform_buffer
-cargo run
-cd ../Eg005.texture_mapping
-cargo run
-cd ../Eg006.box
-cargo run
-
